more work on commaster
