+++ code/modules/computer4/data/terminal/medtrak/medtrak_menu_commands.dm
@@ -12,6 +12,7 @@
 
 /datum/shell_command/medtrak/home/help/exec(datum/c4_file/terminal_program/operating_system/thinkdos/system, datum/c4_file/terminal_program/program, list/arguments, list/options)
 	var/datum/c4_file/terminal_program/medtrak/medtrak = program
+	noop(medtrak)
 	var/list/all_commands = medtrak.home_commands
 	var/list/output = generate_help(system, program, arguments, all_commands)
 	if(output)
