+++ daedalus.dme
@@ -2438,7 +2438,6 @@
 #include "code\modules\antagonists\traitor\objectives\assassination.dm"
 #include "code\modules\antagonists\traitor\objectives\bug_room.dm"
 #include "code\modules\antagonists\traitor\objectives\destroy_item.dm"
-#include "code\modules\antagonists\traitor\objectives\hack_comm_console.dm"
 #include "code\modules\antagonists\traitor\objectives\kill_pet.dm"
 #include "code\modules\antagonists\traitor\objectives\smuggling.dm"
 #include "code\modules\antagonists\traitor\objectives\steal.dm"
